[Refactor] ChatCollectionView를 ChattingViewController로 통합함
